refactor: [25] Form1責務分割とnested class解消 (#53)
* refactor: Form1を責務別partialへ分割し内部クラスを外出し

* fix: ID正規化と売上CSV正規化ロジックを強化
