+++ Sales Management App/Sales Management App/SaleProductOption.cs
@@ -1,0 +1,14 @@
+﻿namespace Sales_Management_App
+{
+    internal class SaleProductOption
+    {
+        public string ProductId { get; set; }
+        public string ProductName { get; set; }
+        public int UnitPrice { get; set; }
+
+        public string DisplayText
+        {
+            get { return string.Format("{0} - {1}", ProductId, ProductName); }
+        }
+    }
+}
